refactor(doctor): keep auth flow reviewable
